+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 * Move initialization of the grid to the template.
 * Added some documentation in the README.
 * Added new form settings to the gridbrowser.
+* Added experimental feature to enable resizing of the grid to use it
+  in responsive layouts.
 
 
 0.1a1